requirements.txt
langchain>=0.3.0
langchain-openai
langchain-chroma
langchain-classic
langchain-community
chromadb
pypdf
fastapi
uvicorn
python-dotenv
pydantic